Remove use of RequiredMixin and change error to warning if non-optional type is missing
RequiredMixin appears to no longer be necessary. Its original purpose was to avoid a runtime error in the core dataclasses library when a base class has a default value for a non-optional field, and a derived class without a default. It appears the dataclasses_json library has made some modifications so that error is not triggered, and instead a warning is emitted.

Also changing the behavior of this library so that warnings are NOT treated as errors when loading GLTF models. Not all models out in the wild strictly adhere to the spec, but I think this library should attempt to be as resilient as possible and still attempt to load the model and emit a warning (instead of just erroring out).